common checks in utils
