+++ validations/scripts/utils/network_checks.py
@@ -14,32 +14,81 @@
 # For any questions about this software or licensing,
 # please email opensource@seagate.com or cortx-questions@seagate.com.
 #
-import os
-import sys
-import logging
-parent_dir_name = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
-sys.path.append(parent_dir_name + "/utils")
 
-from common import run_subprocess_cmd  # noqa: E402
+import logging
+
+from .pillar_get import PillarGet  # noqa: E402
+from .network_connectivity_checks import NetworkValidations
+
 logger = logging.getLogger(__name__)
 
 
-class NetworkValidations():
+class NetworkChecks():
 
     @staticmethod
-    def check_ping(ip):
-        """Check if IP's are reachable"""
-        cmd = f"ping -c 1 {ip}"
-        response = run_subprocess_cmd(cmd)
-        if response[0] == 0:
-            message = f"{ip} is reachable"
-        elif response[0] == 1:
-            message = f"{ip}: Destination Host Unreachable"
-        elif response[0] == 2:
-            message = f"{ip}: Name or service not known"
+    def verify_mgmt_vip():
+        '''Validations for mgmt_vip'''
+        res = PillarGet.get_pillar("cluster:mgmt_vip")
+        if res['ret_code']:
+            res['message'] = "MGMT VIP is not set in pillars"
         else:
-            message = f"{ip}: Not reachable"
-        return {"ret_code": response[0],
-                "response": response[1],
-                "error_msg": response[2],
-                "message": message}
+            res = NetworkValidations.check_ping(res['response'])
+            if res['ret_code']:
+                res['message'] = "Ping to MGMT VIP failed"
+            else:
+                res['message'] = "MGMT VIP is configured"
+        return res
+
+    @staticmethod
+    def verify_cluster_ip():
+        '''Validations for cluster_ip'''
+        res = PillarGet.get_pillar("cluster:cluster_ip")
+        if res['ret_code']:
+            res['message'] = "Cluster IP is not set in pillars"
+        else:
+            res = NetworkValidations.check_ping(res['response'])
+            if res['ret_code']:
+                res['message'] = "Ping to Cluster IP failed"
+            else:
+                res['message'] = "Cluster IP is configured"
+        return res
+
+    @staticmethod
+    def verify_public_data_ip():
+        '''Validations for public data ip'''
+        res = PillarGet.get_pillar("cluster:node_list")
+        nodes = []
+        if not res['ret_code']:
+            nodes = res['response']
+        else:
+            return res
+        for node in nodes:
+            result = PillarGet.get_pillar(f"cluster:{node}:network:data_nw:public_ip_addr")
+            if result['ret_code']:
+                result['message'] = f"Public data IP for {node} is not set in pillars"
+                return result 
+            result = NetworkValidations.check_ping(result['response'])
+            if result['ret_code']:
+                return result
+        result['message'] = "Public data IP is configured"
+        return result
+
+    @staticmethod
+    def verify_private_data_ip():
+        '''Validations for private data ip'''
+        res = PillarGet.get_pillar("cluster:node_list")
+        nodes = []
+        if not res['ret_code']:
+            nodes = res['response']
+        else:
+            return res
+        for node in nodes:
+            result = PillarGet.get_pillar(f"cluster:{node}:network:data_nw:pvt_ip_addr")
+            if result['ret_code']:
+                result['message'] = f"Private data IP for {node} is not set in pillars"
+                return result
+            result = NetworkValidations.check_ping(result['response'])
+            if result['ret_code']:
+                return result
+        result['message'] = "Private data IP is configured"
+        return result
